Replaced whitelist and blacklist terms with allowlist and denylist.
This change is not for ideological reasons, since I do not share the perception of the negative connotations of the previous terms. Simply the next version of the project introduces compatibility breaking changes anyway and I took this opportunity to make this change because the new terms seem to me to reflect the concept better.